Breaking Page into Components
